<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #47572

Problem Summary:

### What is changed and how it works?

add more logs to external backend, disable parallel sort to speed up encode&sort

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)

with a case of 150g, import takes 58m after vs 70m before, mainly speed up encode&sort step. and CPU curve became more smooth.
after
![image](https://github.com/pingcap/tidb/assets/3312245/9469b905-7b37-4de4-b595-2de23180d1d5)
before
![image](https://github.com/pingcap/tidb/assets/3312245/d8963b15-7f45-4e25-8cd2-9ae74d05bf44)

flush log looks like this:
```
[2023/10/16 16:27:57.768 +00:00] [INFO] [writer.go:445] ["flush kv"] [bytes=268339017] [kv-cnt=18160] [stat-size=42504] [sort-time=250.47596ms] [write-time=1.781394854s] [sort-speed(kv/s)=72.5kB] [write-speed(bytes/s)=143.7MiB] [writerID=data/56625710-9837-4c0e-a3f1-b35490f9a008]
[2023/10/16 16:27:57.963 +00:00] [INFO] [writer.go:445] ["flush kv"] [bytes=268339452] [kv-cnt=18160] [stat-size=42504] [sort-time=53.950553ms] [write-time=1.635531465s] [sort-speed(kv/s)=336.6kB] [write-speed(bytes/s)=156.5MiB] [writerID=data/22a1c4ea-b6aa-4a39-9896-b4a461578107]
```

- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->  

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```